requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-tinymce==4.1.0
django-jazzmin==3.0.1
cloudinary==1.42.2
django-cloudinary-storage==0.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1  # For .env support
whitenoise
dj-database-url
Pillow


